搭建前端框架,实现一个简单的MD5登录验证

+++ src/main/java/com/lele/seckill_shop/dao/UserDao.java
@@ -8,5 +8,5 @@ import org.apache.ibatis.annotations.Select;
 public interface UserDao {
 
     @Select("select * from user where id = #{id}")
-    User getById(int id);
+    User getById(Long id);
 }
